+++ packages/terra-dropdown-button/src/terra-dev-site/test/dropdown-button/DefaultDropdownButton.test.jsx
@@ -1,16 +1,26 @@
-import React from 'react';
+import React, { useState } from 'react';
 import classnames from 'classnames/bind';
 import DropdownButton, { Item } from '../../../DropdownButton';
 import styles from './ExtraSpacing.module.scss';
 
 const cx = classnames.bind(styles);
 
-export default () => (
-  <div className={cx('container-spacing-wrapper')}>
-    <DropdownButton label="Dropdown" id="dropdown">
-      <Item label="1st" metaData={{ key: '1st Option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-      <Item label="2nd" metaData={{ key: '2nd Option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-      <Item label="3rd" metaData={{ key: '3rd Option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-    </DropdownButton>
-  </div>
-);
+const DefaultDropdownButton = () => {
+  const [message, setMessage] = useState(' No option clicked');
+
+  return (
+    <div className={cx('container-spacing-wrapper')}>
+      <p>
+        MetaData of :
+        {message}
+      </p>
+      <DropdownButton label="Dropdown" id="dropdown">
+        <Item id="opt1" label="1st" metaData={{ key: '1st Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+        <Item id="opt2" label="2nd" metaData={{ key: '2nd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+        <Item id="opt3" label="3rd" metaData={{ key: '3rd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+      </DropdownButton>
+    </div>
+  );
+};
+
+export default DefaultDropdownButton;