+++ packages/terra-dropdown-button/src/terra-dev-site/test/dropdown-button/DefaultSplitButton.test.jsx
@@ -1,16 +1,27 @@
-import React from 'react';
+import React, { useState } from 'react';
 import classnames from 'classnames/bind';
 import SplitButton, { Item } from '../../../SplitButton';
 import styles from './ExtraSpacing.module.scss';
 
 const cx = classnames.bind(styles);
 
-export default () => (
-  <div className={cx('container-spacing-wrapper')}>
-    <SplitButton primaryOptionLabel="Split" metaData={{ key: 'primary-button' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} id="split">
-      <Item label="1st" metaData={{ key: '1st-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-      <Item label="2nd" metaData={{ key: '2nd-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-      <Item label="3rd" metaData={{ key: '3rd-option' }} onSelect={(event, metaData) => { console.log('Metdata of :', metaData.key); }} />
-    </SplitButton>
-  </div>
-);
+const DefaultSplitButton = () => {
+  const [message, setMessage] = useState(' No option clicked');
+
+  return (
+    <div className={cx('container-spacing-wrapper')}>
+      <p>
+        MetaData of :
+        {message}
+      </p>
+      <SplitButton primaryOptionLabel="Split" metaData={{ key: 'primary-button' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} id="split">
+        <Item id="opt1" label="1st" metaData={{ key: '1st Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+        <Item id="opt2" label="2nd" metaData={{ key: '2nd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+        <Item id="opt3" label="3rd" metaData={{ key: '3rd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
+      </SplitButton>
+    </div>
+  );
+};
+
+export default DefaultSplitButton;
+